Refactored Player Rendering Structure
No function differences in user-visible behavior. Before, there was
sloppy internal structure of player rendering. Now, the player
information and behavior is encapulated and bundled in Player class.

Hierarchy was also added to the folder system. With the addition of the
Player class comes the Sprite class which holds vital rendering
informaiton such as the bounding box of the sprite as well as position.